refactor(ai): SSRF security policy を providerCatalog から分離
PBI 06 (architecture deepening 2026-09-03b, RICE 2.7):

- isAllowedProviderBaseUrl(124行)を providerSecurityPolicy.ts に分離
- providerCatalog は provider data table のみに単純化し、
  security policy は re-export で後方互換維持
- 直近 7 コミットの churn hotspot（security 修正のたびに data
  catalog に触れる問題）を構造的に解消
- policy は catalog map なしで単体テスト可能

既存テスト: providerCatalog 38 + ollamaOriginRule + cspValidator
計 61 tests 全 green
